feat: migrated editor to a component + added routess
